Refactored StudentPanel to inherit from BasePanel.
The BasePanel class now implements View, and StudentPanel is not a subclass of JPanel anymore (a protected reference to a JPanel is installed in BasePanel).
StudentPanel now only defines the right ActionListeners and place them in the buttons.

StudentPanelTest now only contains tests regarding the behaviour of the class when a button is clicked (essentially, tests for ActionListeners)
The remaining tests are done at super class level.